images for products

+++ src/products/products.entity.ts
@@ -1,3 +1,5 @@
+/* eslint-disable */
+import { Images } from './../images/images.entity';
 import { Entity, PrimaryGeneratedColumn, Column, OneToMany } from 'typeorm';
 @Entity()
 export class Products {
@@ -9,13 +11,19 @@ export class Products {
 
   @Column()
   category_id: number;
+  // @ManyToOne(() => Categories, (categories) => categories.category_id)
+  // images: Images[];
 
-  @Column('tinyint')
+  @Column('int')
   price: number;
 
   @Column('int')
   stock: number;
-
+  
   @Column({ length: 1000, nullable: true })
   description: string;
+  
+  // @OneToMany(() => Images, (images) => images.product_id)
+  // images: Images[];
+  
 }
